requirements.txt
crewai
crewai-tools
yfinance
python-dotenv
requests
pandas-ta
streamlit
reportlab